Added Kafka

+++ auth-service/src/main/java/com/prevpaper/auth/services/Impl/AuthServiceImpl.java
@@ -1,9 +1,7 @@
 package com.prevpaper.auth.services.Impl;
 
 import com.prevpaper.auth.config.JwtService;
-import com.prevpaper.auth.dto.LoginRequest;
-import com.prevpaper.auth.dto.SignupRequest;
-import com.prevpaper.auth.dto.VerifyOtpRequest;
+import com.prevpaper.auth.dto.*;
 import com.prevpaper.auth.entities.Session;
 import com.prevpaper.auth.entities.User;
 import com.prevpaper.auth.entities.Role;
@@ -14,16 +12,14 @@ import com.prevpaper.auth.repositories.RoleRepository;
 import com.prevpaper.auth.repositories.VerificationTokenRepository;
 import com.prevpaper.auth.services.AuditService;
 import com.prevpaper.auth.services.AuthService;
+import com.prevpaper.auth.utils.GererateOtp;
 import com.prevpaper.comman.dto.ApiResponse;
-import com.prevpaper.comman.enums.AccountStatus;
-import com.prevpaper.comman.enums.AuthProvider;
-import com.prevpaper.comman.enums.TokenType;
-import com.prevpaper.comman.enums.UserRole;
+import com.prevpaper.comman.dto.CommonNotificationRequest;
+import com.prevpaper.comman.enums.*;
 
 // CORRECTED IMPORTS: Use jakarta instead of javax
-import com.prevpaper.comman.exception.InvalidOtpException;
-import com.prevpaper.comman.exception.ResourceNotFoundException;
-import com.prevpaper.comman.exception.RoleExceptionHandler;
+import com.prevpaper.comman.exception.*;
+import com.prevpaper.comman.producer.NotificationProducer;
 import jakarta.servlet.http.Cookie;
 //import jakarta.servlet.http.HttpServletRequest;
 //import jakarta.servlet.http.HttpServletResponse;
@@ -35,13 +31,12 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionSynchronization;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
+import java.util.*;
+
 @Service
 @RequiredArgsConstructor
 public class AuthServiceImpl implements AuthService {
@@ -50,9 +45,12 @@ public class AuthServiceImpl implements AuthService {
     private final RoleRepository roleRepository;
     private final JwtService jwtService;
     private final SessionRepository sessionRepository;
-    private final PasswordEncoder passwordEncoder;
     private final AuditService auditService;
     private  final VerificationTokenRepository verificationTokenRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final NotificationProducer producer;
+
+
     @Value("${jwt.refresh-expiration}")
     private long refreshExpiration;
 
@@ -60,25 +58,23 @@ public class AuthServiceImpl implements AuthService {
     @Override
     @Transactional
     public ApiResponse<Map<String, String>> registerUser(SignupRequest request, HttpServletRequest req) {
-        // 1. Validation
+
         if ((request.getEmail() == null || request.getEmail().isBlank()) &&
                 (request.getPhoneNumber() == null || request.getPhoneNumber().isBlank())) {
             return new ApiResponse<>(false, "Email or Phone Number is required", null, System.currentTimeMillis());
         }
-
-        // 2. Duplicate Check
+        System.out.println("UPto here all things are good");
         if (request.getEmail() != null && userRepository.existsByEmail(request.getEmail())) {
             return new ApiResponse<>(false, "Email already exists", null, System.currentTimeMillis());
         }
         if (request.getPhoneNumber() != null && userRepository.existsByPhoneNumber(request.getPhoneNumber())) {
             return new ApiResponse<>(false, "Phone number already exists", null, System.currentTimeMillis());
         }
+        System.out.println("UPto here all things are good1");
 
-        // 3. Fetch default role
         Role studentRole = roleRepository.findByRoleName(UserRole.STUDENT)
                 .orElseThrow(() -> new RoleExceptionHandler("Default Role not found"));
 
-        // 4. Create User
         User user = User.builder()
                 .email(request.getEmail())
                 .phoneNumber(request.getPhoneNumber())
@@ -91,32 +87,41 @@ public class AuthServiceImpl implements AuthService {
                 .build();
 
         userRepository.save(user);
-
-
-
-        // 5. Generate Tokens
+        System.out.println("UPto here all things are good2");
         String accessToken = jwtService.generateAccessToken(user);
         String refreshToken = jwtService.generateRefreshToken(user);
 
-        // NEW: Generate 6-digit OTP
-        String otp = String.valueOf((int) (Math.random() * 900000) + 100000);
-
-        // send otp
-
+        String otp = GererateOtp.getOTP();
+        String hashedOtp = passwordEncoder.encode(otp);
+        System.out.println("UPto here all things are good3");
         verificationTokenRepository.save(
                 VerificationToken.builder()
-                        .token(otp)
+                        .token(hashedOtp)
                         .user(user)
                         .type(TokenType.EMAIL_VERIFY)
-                        .expiryDate(LocalDateTime.now().plusMinutes(10)) // 10 min expiry
+                        .expiryDate(LocalDateTime.now().plusMinutes(10))
                         .build()
         );
+        System.out.println("UPto here all things are good4");
+        CommonNotificationRequest notificationRequest = CommonNotificationRequest.builder()
+                .userId(user.getUserId())
+                .recipient(request.getEmail() != null ? request.getEmail() : request.getPhoneNumber())
+                .title("OTP Verification")
+                .message("Your verification OTP is " + otp + ". It expires in 10 minutes.")
+                .notificationTypes(List.of(NotificationType.EMAIL)) // for channel
+                .eventType(com.prevpaper.comman.enums.NotificationEventType.OTP_SENT)        // business event
+                .build();
 
-        // todo for send the otp
+        System.out.println(notificationRequest);
 
-        // if otp is verify successfully we will generate token and send to frontend otherwise not
-
-        // 6. Store Session
+        TransactionSynchronizationManager.registerSynchronization(
+                new TransactionSynchronization() {
+                    @Override
+                    public void afterCommit() {
+                        producer.sendOtpNotification(user.getUserId().toString(), notificationRequest);
+                    }
+                }
+        );
 
         sessionRepository.save(
                 Session.builder()
@@ -127,18 +132,14 @@ public class AuthServiceImpl implements AuthService {
                         .build()
         );
 
-        // 7. Prepare tokens
         Map<String, String> tokens = new HashMap<>();
         tokens.put("accessToken", accessToken);
         tokens.put("refreshToken", refreshToken);
 
-        // 8. Audit log
         auditService.logAction(user, "SIGNUP", user.getUserId().toString(), req);
 
-        // 9. Return response
         return new ApiResponse<>(true, "User registered successfully", tokens, System.currentTimeMillis());
     }
-
     // Login Method
     @Override
     public ApiResponse<Map<String, String>> loginUser(LoginRequest loginRequest,
@@ -198,12 +199,10 @@ public class AuthServiceImpl implements AuthService {
         User user = userRepository.findByEmail(request.getEmail())
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
-        VerificationToken otpToken = verificationTokenRepository
-                .findByUser(user)
+        VerificationToken otpToken = verificationTokenRepository.findByUserAndType(user, TokenType.EMAIL_VERIFY)
                 .orElseThrow(() -> new InvalidOtpException("OTP not found"));
 
-        System.out.println("User found"+ user);
-        System.out.println("Verification TOkan :" + otpToken);
+
 
 
         // Check OTP match
@@ -213,19 +212,110 @@ public class AuthServiceImpl implements AuthService {
 
         // Check expiry
         if (otpToken.getExpiryDate().isBefore(LocalDateTime.now())) {
+            verificationTokenRepository.delete(otpToken); // delete expired OTP
             throw new InvalidOtpException("OTP expired");
         }
-
         // Activate account
         user.setAccountStatus(AccountStatus.ACTIVE);
         userRepository.save(user);
 
-        // Delete OTP after successful verification
-        verificationTokenRepository.delete(otpToken);
+        otpToken.setVerified(true);
+        verificationTokenRepository.save(otpToken);
+
 
         return ApiResponse.success(
                 "OTP verified successfully",
                 Map.of("verified", "true")
+        );
+    }
+
+
+
+    @Override
+    public ApiResponse<Map<String, String>> forgotPassword(ForgotPasswordRequest request, HttpServletRequest httpRequest) {
+
+        if(request.getEmail() == null || request.getEmail().isBlank()){
+            throw new EmptyInputBoxException("Email is Required");
+        }
+
+        String email = request.getEmail().trim();
+
+
+        Optional<User> userOptional = userRepository.findByEmail(email);
+
+        if(userOptional.isEmpty()){
+            return ApiResponse.success(
+                    "If the email exists, OTP has been sent",
+                    Map.of("success","true")
+            );
+        }
+
+        User user = userOptional.get();
+
+        verificationTokenRepository.deleteByUserAndType(user, TokenType.PASSWORD_RESET);
+
+        String otp = GererateOtp.getOTP();
+
+        verificationTokenRepository.save(
+                VerificationToken.builder()
+                        .token(otp)
+                        .user(user)
+                        .type(TokenType.PASSWORD_RESET)
+                        .expiryDate(LocalDateTime.now().plusMinutes(10)) // 10 min expiry
+                        .build()
+        );
+
+
+
+        // send otp by notification service
+        return ApiResponse.success(
+                "OTP Sent successfully",
+                Map.of("send OTP", "true")
+        );
+
+
+    }
+
+
+
+    @Override
+    public ApiResponse<Map<String, String>> resetPassword(
+            ResetPasswordRequest request,
+            HttpServletRequest httpRequest) {
+
+        if(request.getEmail() == null || request.getEmail().isBlank() ||
+                request.getNewPassword() == null || request.getNewPassword().isBlank() ||
+                request.getConfirmPassword() == null || request.getConfirmPassword().isBlank()) {
+            throw new EmptyInputBoxException("All fields are required");
+        }
+
+        if(!request.getNewPassword().equals(request.getConfirmPassword())){
+            throw new WrongCredentialsException("Passwords do not match");
+        }
+
+        String email = request.getEmail().trim();
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+
+        VerificationToken otpToken = verificationTokenRepository
+                .findByUserAndTypeAndVerified(user, TokenType.PASSWORD_RESET, true)
+                .orElseThrow(() -> new InvalidOtpException("OTP not verified or expired"));
+
+        if (otpToken.getExpiryDate().isBefore(LocalDateTime.now())) {
+            verificationTokenRepository.delete(otpToken);
+            throw new InvalidOtpException("OTP expired");
+        }
+
+        String hashedPassword = passwordEncoder.encode(request.getNewPassword());
+        user.setPasswordHash(hashedPassword);
+        userRepository.save(user);
+
+        verificationTokenRepository.delete(otpToken);
+
+        return ApiResponse.success(
+                "Password reset successfully",
+                Map.of("passwordReset", "true")
         );
     }
 }
